Refactoring panda.py and obsoleting fastCGI configurations
