+++ pandaserver/server/allowed_methods.py
@@ -1,0 +1,147 @@
+# List of methods that can be executed by the clients
+
+allowed_methods = []
+
+# methods from pandaserver.taskbuffer.Utils
+allowed_methods += [
+    "isAlive",
+    "putFile",
+    "deleteFile",
+    "getServer",
+    "updateLog",
+    "fetchLog",
+    "touchFile",
+    "getVomsAttr",
+    "putEventPickingRequest",
+    "getAttr",
+    "uploadLog",
+    "put_checkpoint",
+    "delete_checkpoint",
+    "put_file_recovery_request",
+    "put_workflow_request",
+]
+
+# methods from pandaserver.dataservice.DataService
+allowed_methods += ["datasetCompleted", "updateFileStatusInDisp"]
+
+# methods from pandaserver.jobdispatcher.JobDispatcher
+allowed_methods += [
+    "getJob",
+    "updateJob",
+    "getStatus",
+    "genPilotToken",
+    "getEventRanges",
+    "updateEventRange",
+    "getKeyPair",
+    "updateEventRanges",
+    "getDNsForS3",
+    "getProxy",
+    "getCommands",
+    "ackCommands",
+    "checkJobStatus",
+    "checkEventsAvailability",
+    "updateJobsInBulk",
+    "getResourceTypes",
+    "updateWorkerPilotStatus",
+    "get_max_worker_id",
+    "get_events_status",
+]
+
+# methods from pandaserver.userinterface.UserIF
+allowed_methods += [
+    "submitJobs",
+    "getJobStatus",
+    "queryPandaIDs",
+    "killJobs",
+    "reassignJobs",
+    "getJobStatistics",
+    "getJobStatisticsPerSite",
+    "resubmitJobs",
+    "queryLastFilesInDataset",
+    "getPandaIDsSite",
+    "getJobsToBeUpdated",
+    "updateProdDBUpdateTimes",
+    "runTaskAssignment",
+    "getAssigningTask",
+    "getSiteSpecs",
+    "getCloudSpecs",
+    "seeCloudTask",
+    "queryJobInfoPerCloud",
+    "registerProxyKey",
+    "getProxyKey",
+    "getJobIDsInTimeRange",
+    "getPandIDsWithJobID",
+    "getFullJobStatus",
+    "getJobStatisticsForBamboo",
+    "getNUserJobs",
+    "addSiteAccess",
+    "listSiteAccess",
+    "getFilesInUseForAnal",
+    "updateSiteAccess",
+    "getPandaClientVer",
+    "getSlimmedFileInfoPandaIDs",
+    "getQueuedAnalJobs",
+    "getHighestPrioJobStat",
+    "getActiveDatasets",
+    "setCloudTaskByUser",
+    "getSerialNumberForGroupJob",
+    "checkMergeGenerationStatus",
+    "getNumPilots",
+    "retryFailedJobsInActive",
+    "getJobStatisticsWithLabel",
+    "getPandaIDwithJobExeID",
+    "getJobStatisticsPerUserSite",
+    "getDisInUseForAnal",
+    "getLFNsInUseForAnal",
+    "getScriptOfflineRunning",
+    "setDebugMode",
+    "insertSandboxFileInfo",
+    "checkSandboxFile",
+    "changeJobPriorities",
+    "insertTaskParams",
+    "killTask",
+    "finishTask",
+    "getJediTasksInTimeRange",
+    "getJediTaskDetails",
+    "retryTask",
+    "getRetryHistory",
+    "changeTaskPriority",
+    "reassignTask",
+    "changeTaskAttributePanda",
+    "pauseTask",
+    "resumeTask",
+    "increaseAttemptNrPanda",
+    "killUnfinishedJobs",
+    "changeTaskSplitRulePanda",
+    "changeTaskModTimePanda",
+    "avalancheTask",
+    "getPandaIDsWithTaskID",
+    "reactivateTask",
+    "getTaskStatus",
+    "reassignShare",
+    "listTasksInShare",
+    "getTaskParamsMap",
+    "updateWorkers",
+    "harvesterIsAlive",
+    "reportWorkerStats",
+    "reportWorkerStats_jobtype",
+    "addHarvesterDialogs",
+    "getJobStatisticsPerSiteResource",
+    "setNumSlotsForWP",
+    "reloadInput",
+    "enableJumboJobs",
+    "updateServiceMetrics",
+    "getUserJobMetadata",
+    "getJumboJobDatasets",
+    "getGShareStatus",
+    "sweepPQ",
+    "get_job_statistics_per_site_label_resource",
+    "relay_idds_command",
+    "send_command_to_job",
+    "execute_idds_workflow_command",
+    "set_user_secret",
+    "get_user_secrets",
+    "get_ban_users",
+    "get_files_in_datasets",
+    "release_task",
+]
